client: API rework: bring the Authorization Token at the Client level
This commit breaks the API and moves the authentication detail away from individual invocations.
The Client (now better documented) has now a SetToken* set of methods which enables setting an
authentication context dedicated to the client and not on a per-call basis.

This improves the UX all along except for what concerns the Device SDK case. However, given that
this interaction will in turn be screened by the SDK, it's an acceptable tradeoff.

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>