validation: Validate nested arrays correctly
When marshaling out of a composite map (e.g.: when dealing with an aggregate), a Go slice might
end up being []interface{} rather than its real type. In such cases, we have to recursively validate
the internal slice element to make something out of it.

This fixes pretty much any possible validation for astartectl send-data when it comes to aggregates.

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>